major refactor; add support for column filtering objects (ColumnFilter and ColumnFilterChain)
